Rearrange folder structure and add setup react-router-dom and active links
