+++ src/App.js
@@ -1,14 +1,19 @@
 import './App.css';
-import Cart from './pages/Home/Cart';
-import Home from './pages/Home/Home';
+import { Routes, Route} from "react-router-dom";
+import Cart from './pages/Cart';
+import Home from './pages/Home';
+
 
 
 
 function App() {
   return (
     <>
-      {/* <Home/> */}
-      <Cart/>
+    <Routes>
+      <Route path="/react-shopeefy-app" element={<Home />}/>
+      <Route path="/react-shopeefy-app/cart" element={<Cart />}/>
+      
+    </Routes>
     </>
   );
 }
