+++ src/App.scss
@@ -4,13 +4,8 @@
 
 
 :root{
-
-    // --MainHeaderheight: 225px;
-    --MainHeaderheight: 80px;
-
-   
+    --serachSectionHight: 150px
 }
-
 
 
 body{
@@ -41,12 +36,12 @@ ul{
    top:0;
    left: 0;
    z-index: 1000;
-   height: var(--MainHeaderheight);
+//    height: var(--MainHeaderheight);
    background: green;
 }
 
 .mainBody{
-    margin-top: var(--MainHeaderheight);
+    margin-top: calc(var.$NavHeight + var(--serachSectionHight));
     background: var.$light-grey-color;
     display: flex;
     flex-wrap: wrap;
@@ -61,15 +56,15 @@ ul{
 }
 
 .cartMain{
-    margin-top: var(--MainHeaderheight);
+    // margin-top: var(--MainHeaderheight);
     background: var.$light-grey-color;
     padding: 1.5em 2em;
 }
 
+
 @include media.desktop1000{
     :root{
-    --MainHeaderheight: 320px;
+        --serachSectionHight: 220px
     }
-}  
-
- 
+    
+}
