+++ src/components/Navbar/Navbar.js
@@ -1,24 +1,24 @@
 import React from 'react'
+import { Link } from 'react-router-dom';
 import {MdShoppingCart} from 'react-icons/md'
 import styles from './Navbar.module.scss';
-
 
 const Navbar = () => {
   return (
     <div className={styles.navBar}>
         <ul className={styles.navBar__menu}>
-            <li className={styles.navBar__logo}><h2><a href="#" className='link'>Shopeefy</a></h2></li>
+            <li className={styles.navBar__logo}><h2><Link to="/react-shopeefy-app" className='link'>Shopeefy</Link></h2></li>
             <li className={styles.navBar__user}>
                 <div className={styles.navBar__user__pic}>
-                    <img className={styles.img} src="./images/profile.jpg" alt="profilepic" />
+                    <img className={styles.img} src="../images/profile.jpg" alt="profilepic" />
                 </div>
                 <div className={styles.navBar__user__name}>Akshay Tambe</div>
             </li>
             <li>
-                <a href='#' className={`${styles.navBar__cart} link`}>
+                <Link to="/react-shopeefy-app/cart" className={`${styles.navBar__cart} link`}>
                 <div className={styles.lable}>1</div> 
                 <MdShoppingCart/>
-                </a>
+                </Link>
             </li>
         </ul>
