+++ src/components/SearchContainer/SearchContainer.module.scss
@@ -8,7 +8,7 @@
     display: flex;
     flex-direction: column;
     align-items: center;
-    height: calc(100% - var.$NavHeight);
+    height: var(--serachSectionHight);
     position: relative;
     justify-content: center;
     background: var.$light-color;
